requirements.txt
asgiref
cachetools
certifi
charset-normalizer
Django
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
numpy
oauthlib
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
python-dotenv
python-json-logger
requests
requests-oauthlib
rsa
scipy
soupsieve
sqlparse
uritemplate
urllib3
django-crontab